Cleanup and docs.
